backend/requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
apify-client==1.6.4
bcrypt==5.0.0
beautifulsoup4==4.14.3
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.5
cymem==2.0.13
deprecation==2.1.0
distro==1.9.0
faiss-cpu>=1.9.0
fastapi==0.115.5
filelock==3.25.2
fsspec==2026.2.0
gotrue==2.12.4
greenlet==3.3.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
huggingface_hub==0.36.2
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
langcodes==3.5.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.15
networkx==3.6.1
numpy==1.26.4
openai==1.55.3
packaging==26.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==12.1.1
pluggy==1.6.0
postgrest==1.1.1
preshed==3.0.13
psycopg2==2.9.11
pycparser==3.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.19.2
PyJWT==2.12.1
PyPDF2==3.0.1
pypdfium2==5.6.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.3
realtime==2.5.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==3.3.1
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.48
srsly==2.5.3
starlette==0.41.3
storage3==0.12.1
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.2
sympy==1.13.1
thinc==8.3.11
threadpoolctl==3.6.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.3
transformers==4.46.3
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.32.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
websockets==15.0.1
wrapt==2.1.2
